requirements.txt
aiohttp==3.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
bitarray==3.0.0
blinker==1.9.0
boto3==1.35.1
botocore==1.35.99
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
cytoolz==1.0.1
distro==1.9.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.2
oauthlib==3.2.2
openai==1.61.0
parsimonious==0.10.0
pgvector==0.3.6
propcache==0.2.1
psycopg2-binary==2.9.10
py-cord==2.6.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.10
pyunormalize==16.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==1.3.1
rlp==4.1.0
s3transfer==0.10.4
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
sniffio==1.3.1
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==1.0.0
tqdm==4.67.1
tweepy==4.14.0
types-requests==2.32.0.20241016
typing_extensions==4.12.2
urllib3==2.3.0
web3==7.6.1
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.3